Fix: refactoring
- moved components inside lib
