+++ src/lib/components/hero.svelte
