+++ src/lib/components/high-level-architecture.svelte
