+++ src/lib/components/index.ts
@@ -1,0 +1,25 @@
+import DropDown from './dropdown.svelte';
+import Hero from './hero.svelte';
+import HighLevelArchitecture from './high-level-architecture.svelte';
+import HowToIntro from './how-to-intro.svelte';
+import Invite from './invite.svelte';
+import NewRepository from './newRepository.svelte';
+import Overview from './overview.svelte';
+import Pulse from './pulse.svelte';
+import Register from './register.svelte';
+import Repository from './repository.svelte';
+import Signin from './signin.svelte';
+import Signup from './signup.svelte';
+
+export { DropDown };
+export { Hero };
+export { HighLevelArchitecture };
+export { HowToIntro };
+export { Invite };
+export { NewRepository };
+export { Overview };
+export { Pulse };
+export { Register };
+export { Repository };
+export { Signin };
+export { Signup };
